Refactor HttpClient usage and add content saving feature
The HttpClient in LeetCodeSnippet has been changed to static and the method GetProblemDetails has been made static as well. Also, the 'buffers' directory is no longer included in the csproj file. A new feature to save problem content retrieved from LeetCode to a buffer directory has been implemented in the Main function of the Program.cs file. The changes provide a streamlined approach for HttpClient usage and enhanced functionality for content retrieval.